# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "servers": [{ "url": "http://localhost:18080", "description": "Generated server url" }],
...
  "paths": {
    "/api/v1/products/{id}": {
      "get": {
        "tags": ["Product"],
        "summary": "商品詳細を取得する",
        "operationId": "findProductById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/ProductDetailResponse" } }
            }
          },
          "404": {
            "description": "見つからない",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/ProductDetailResponse" } }
            }
          }
        }
      },
      "put": {
        "tags": ["Product"],
        "summary": "商品を更新する",
        "operationId": "updateProduct",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProductRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "更新成功" } }
      },
      "delete": {
        "tags": ["Product"],
        "summary": "商品を削除する",
        "operationId": "deleteProduct",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VersionRequest" } }
          },
          "required": true
        },
        "responses": { "204": { "description": "削除成功" } }
      }
    },
...
    "/api/v1/categories/{id}": {
      "get": {
        "tags": ["Category"],
        "summary": "カテゴリ詳細を取得する",
        "operationId": "findCategoryById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/CategoryDetailResponse" } }
            }
          },
          "404": {
            "description": "見つからない",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/CategoryDetailResponse" } }
            }
          }
        }
      },
      "put": {
        "tags": ["Category"],
        "summary": "カテゴリを更新する",
        "operationId": "updateCategory",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCategoryRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "更新成功" } }
      },
      "delete": {
        "tags": ["Category"],
        "summary": "カテゴリを削除する",
        "operationId": "deleteCategory",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/DeleteRequest" } }
          },
          "required": true
        },
        "responses": { "204": { "description": "削除成功" } }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": ["Product"],
        "summary": "商品一覧を取得する",
        "operationId": "listProduct",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          { "name": "status", "in": "query", "required": false, "schema": { "type": "string" } },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": { "type": "integer", "default": 0, "minimum": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": { "type": "integer", "default": 20, "minimum": 1 }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/PageProductSummaryResponse" } }
            }
          }
        }
      },
      "post": {
        "tags": ["Product"],
        "summary": "商品を作成する",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProductRequest" }
            }
          },
          "required": true
        },
        "responses": { "201": { "description": "作成成功" } }
      }
    },
...
    "/api/v1/categories": {
      "get": {
        "tags": ["Category"],
        "summary": "ルートカテゴリ一覧を取得する",
        "operationId": "listCategory",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": { "type": "integer", "default": 0, "minimum": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": { "type": "integer", "default": 20, "minimum": 1 }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "*/*": { "schema": { "$ref": "#/components/schemas/PageCategorySummaryResponse" } }
            }
          }
        }
      },
      "post": {
        "tags": ["Category"],
        "summary": "カテゴリを作成する",
        "operationId": "createCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCategoryRequest" }
            }
          },
          "required": true
        },
        "responses": { "201": { "description": "作成成功" } }
      }
    },
    "/api/v1/products/{id}/unpublish": {
      "patch": {
        "tags": ["Product"],
        "summary": "商品を非公開にする",
        "operationId": "unpublishProduct",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VersionRequest" } }
          },
          "required": true
        },
        "responses": { "200": { "description": "非公開成功" } }
      }
    },
    "/api/v1/products/{id}/publish": {
      "patch": {
        "tags": ["Product"],
        "summary": "商品を公開する",
        "operationId": "publishProduct",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VersionRequest" } }
          },
          "required": true
        },
        "responses": { "200": { "description": "公開成功" } }
      }
    },
    "/api/v1/products/{id}/archive": {
      "patch": {
        "tags": ["Product"],
        "summary": "商品をアーカイブする",
        "operationId": "archiveProduct",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VersionRequest" } }
          },
          "required": true
        },
        "responses": { "200": { "description": "アーカイブ成功" } }
      }
    },
    "/api/v1/categories/{id}/move": {
      "patch": {
        "tags": ["Category"],
        "summary": "カテゴリを移動する",
        "operationId": "moveCategory",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/MoveCategoryRequest" } }
          },
          "required": true
        },
        "responses": { "200": { "description": "移動成功" } }
      }
    },
    "/api/v1/categories/{id}/children": {
      "get": {
        "tags": ["Category"],
        "summary": "直接の子カテゴリ一覧を取得する",
        "operationId": "findCategoryChildren",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CategorySummaryResponse" }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "description": { "type": "string", "minLength": 1 },
          "categoryId": { "type": "string", "minLength": 1 },
          "version": { "type": "integer", "format": "int32" }
        },
        "required": ["categoryId", "description", "name"]
      },
...
      "UpdateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 50, "minLength": 0 },
          "sortOrder": { "type": "integer", "format": "int32" },
          "version": { "type": "integer", "format": "int32" }
        },
        "required": ["name"]
      },
      "CreateProductRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "description": { "type": "string", "minLength": 1 },
          "categoryId": { "type": "string", "minLength": 1 },
          "sku": { "type": "string", "minLength": 1 }
        },
        "required": ["categoryId", "description", "name", "sku"]
      },
...
      "CreateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 50, "minLength": 0 },
          "sortOrder": { "type": "integer", "format": "int32" },
          "parentCategoryId": { "type": "string" }
        },
        "required": ["name"]
      },
      "VersionRequest": {
        "type": "object",
        "properties": { "version": { "type": "integer", "format": "int32" } }
      },
      "MoveCategoryRequest": {
        "type": "object",
        "properties": {
          "newParentCategoryId": { "type": "string" },
          "version": { "type": "integer", "format": "int32" }
        }
      },
      "PageProductSummaryResponse": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProductSummaryResponse" }
          },
          "number": { "type": "integer", "format": "int32" },
          "pageable": { "$ref": "#/components/schemas/PageableObject" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "numberOfElements": { "type": "integer", "format": "int32" },
          "empty": { "type": "boolean" }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": { "type": "integer", "format": "int64" },
          "paged": { "type": "boolean" },
          "pageNumber": { "type": "integer", "format": "int32" },
          "pageSize": { "type": "integer", "format": "int32" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "unpaged": { "type": "boolean" }
        }
      },
      "ProductSummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "status": { "type": "string" },
          "categoryId": { "type": "string" }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": { "type": "boolean" },
          "sorted": { "type": "boolean" },
          "unsorted": { "type": "boolean" }
        }
      },
      "ProductDetailResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "categoryId": { "type": "string" },
          "sku": { "type": "string" },
          "status": { "type": "string" },
          "version": { "type": "integer", "format": "int32" }
        }
      },
...
      "CategorySummaryResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "sortOrder": { "type": "integer", "format": "int32" }
        }
      },
      "PageCategorySummaryResponse": {
        "type": "object",
        "properties": {
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CategorySummaryResponse" }
          },
          "number": { "type": "integer", "format": "int32" },
          "pageable": { "$ref": "#/components/schemas/PageableObject" },
          "sort": { "$ref": "#/components/schemas/SortObject" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "numberOfElements": { "type": "integer", "format": "int32" },
          "empty": { "type": "boolean" }
        }
      },
      "AncestorResponse": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" } }
      },
      "CategoryDetailResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "sortOrder": { "type": "integer", "format": "int32" },
          "parentCategoryId": { "type": "string" },
          "version": { "type": "integer", "format": "int32" },
          "ancestors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AncestorResponse" }
          }
        }
      },
      "DeleteRequest": {
        "type": "object",
        "properties": { "version": { "type": "integer", "format": "int32" } }
      }
    }
  }
}